backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
gunicorn==21.2.0
schedule==1.2.0
azure-cognitiveservices-speech==1.34.0
datasets==2.15.0
phonemizer==3.2.1
librosa==0.10.1
soundfile==0.12.1
jiwer==3.0.3
